fix(sip): persist dial-in information from signaling settings in the store
- optionally rendered tab component is not reliable for that
